+++ src/world/Level0.cpp
@@ -417,7 +417,7 @@ bool ChatHandler::HandleRatingsCommand(const char* args, WorldSession* m_session
     if (m_plyr == NULL)
         return false;
 
-    for (uint32 i = 0; i < 24; i++)
+    for (uint8 i = 0; i < 24; i++)
     {
         m_plyr->ModUnsigned32Value(PLAYER_FIELD_COMBAT_RATING_1 + i, i);
     }
