+++ src/world/Object.cpp
@@ -679,7 +679,7 @@ void Object::_BuildValuesUpdate(ByteBuffer* data, UpdateMask* updateMask, Player
                         {
                             if (qle->GetQuest()->count_required_mob == 0)
                                 continue;
-                            for (uint32 i = 0; i < 4; ++i)
+                            for (uint8 i = 0; i < 4; ++i)
                             {
                                 if (qle->GetQuest()->required_mob[i] == static_cast<int32>(go->GetEntry()) && qle->GetMobCount(i) < qle->GetQuest()->required_mobcount[i])
                                 {
